Use uint32 for file index.

+++ include/evb/bu/FileHandler.h
@@ -68,8 +68,8 @@ namespace evb {
       void defineJSON(const boost::filesystem::path&) const;
       void calcAdler32(const unsigned char* address, size_t length);
 
-      typedef std::map<uint32_t,uint16_t> LumiIndex;
-      static uint16_t getNextIndex(const uint32_t lumiSection);
+      typedef std::map<uint32_t,uint32_t> LumiIndex;
+      static uint32_t getNextIndex(const uint32_t lumiSection);
       static boost::mutex indexMutex_;
       static LumiIndex lumiIndex_;
       static uint32_t lastLumiSection_;
